major movies helper refactor
